requirements.txt
bring-api>=6.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
